requirements.txt
requests
numpy
mypy
black
flake8
pylint
Cython
alpaca-trade-api>=0.25
sklearn
google-cloud-logging
therapist
jupyter
ta-lib
influxdb
asyncpg
psycopg2-binary
sqlalchemy
pygit2
polygon-api-client
google-cloud-error-reporting
